# Tolk v1.0: lazy loading, AST inlining, peephole optimizations, TVM 11