--- a/crypto/smartcont/tolk-stdlib/gas-payments.tolk
+++ b/crypto/smartcont/tolk-stdlib/gas-payments.tolk
@@ -45,16 +45,17 @@ fun calculateGasFeeWithoutFlatPrice(workchain: int8, gasUsed: coins): coins
 fun calculateStorageFee(workchain: int8, seconds: int, bits: int, cells: int): coins
     asm(cells bits seconds workchain) "GETSTORAGEFEE";
 
-/// Calculates amount of nanotoncoins you should pay to send a message of specified size.
-fun calculateMessageFee(workchain: int8, bits: int, cells: int): coins
+/// Calculates amount of nanotoncoins you should pay to send a message of a specified size.
+fun calculateForwardFee(workchain: int8, bits: int, cells: int): coins
     asm(cells bits workchain) "GETFORWARDFEE";
 
-/// Same as [calculateMessageFee], but without lump price (you have supposed to read https://docs.ton.org/develop/howto/fees-low-level)
-fun calculateMessageFeeWithoutLumpPrice(workchain: int8, bits: int, cells: int): coins
+/// Same as [calculateForwardFee], but without lump price (you have supposed to read https://docs.ton.org/develop/howto/fees-low-level)
+fun calculateForwardFeeWithoutLumpPrice(workchain: int8, bits: int, cells: int): coins
     asm(cells bits workchain) "GETFORWARDFEESIMPLE";
 
 /// Calculates fee that was paid by the sender of an incoming internal message.
-fun calculateOriginalMessageFee(workchain: int8, incomingFwdFee: coins): coins
+@deprecated("use modern `onInternalMessage` and access `in.originalForwardFee` directly")
+fun calculateOriginalForwardFee(workchain: int8, incomingFwdFee: coins): coins
     asm(incomingFwdFee workchain) "GETORIGINALFWDFEE";
 
 /// Returns the amount of nanotoncoins current contract debts for storage. ("due" and "debt" are synonyms)