--- a/crypto/smartcont/tolk-stdlib/lisp-lists.tolk
+++ b/crypto/smartcont/tolk-stdlib/lisp-lists.tolk
@@ -9,25 +9,25 @@ tolk 0.99
 
 @pure
 fun createEmptyList(): tuple
-    asm "PUSHNULL";
+    asm "PUSHNULL"
 
 /// Adds an element to the beginning of lisp-style list.
 /// Note, that it does not mutate the list: instead, it returns a new one (it's a lisp pattern).
 @pure
 fun listPrepend<X>(head: X, tail: tuple?): tuple
-    asm "CONS";
+    asm "CONS"
 
 /// Extracts the head and the tail of lisp-style list.
 @pure
 fun listSplit<X>(list: tuple): (X, tuple?)
-    asm "UNCONS";
+    asm "UNCONS"
 
 /// Returns the head of lisp-style list.
 @pure
 fun listGetHead<X>(list: tuple): X
-    asm "CAR";
+    asm "CAR"
 
 /// Returns the tail of lisp-style list.
 @pure
 fun listGetTail(list: tuple): tuple?
-    asm "CDR";
+    asm "CDR"