--- a/tolk/asmops.cpp
+++ b/tolk/asmops.cpp
@@ -223,6 +223,8 @@ AsmOp AsmOp::Parse(SrcLocation loc, const std::string& custom_op) {
     return AsmOp::Push(loc, 0);
   } else if (custom_op == "OVER") {
     return AsmOp::Push(loc, 1);
+  } else if (custom_op.ends_with(" PUSHINT") && custom_op[0] >= '1' && custom_op[0] <= '9' && custom_op.find(' ') == custom_op.rfind(' ')) {
+    return AsmOp::IntConst(loc, td::string_to_int256(custom_op.substr(0, custom_op.find(' '))));
   } else {
     return AsmOp::Custom(loc, custom_op);
   }