--- a/tolk/ast-aux-data.h
+++ b/tolk/ast-aux-data.h
@@ -17,6 +17,7 @@
 #pragma once
 
 #include "ast.h"
+#include "lazy-helpers.h"
 
 /*
  *   This file contains a schema of aux_data inside ast_artificial_aux_vertex
@@ -32,6 +33,29 @@ struct AuxData_ForceFiftLocation final : ASTAuxData {
 
   explicit AuxData_ForceFiftLocation(SrcLocation forced_loc)
     : forced_loc(forced_loc) {
+  }
+};
+
+// AuxData_LazyObjectLoadFields is a special auto-inserted vertex to load fields of a lazy struct;
+// example: `var p = lazy Point.fromSlice(s); aux "load x"; return p.x`
+struct AuxData_LazyObjectLoadFields final : ASTAuxData {
+  LocalVarPtr var_ref;              // comes from `lazy`
+  TypePtr union_variant;            // not just `o` but `match(o) { V1 => here }`
+  StructFieldPtr field_ref;         // not just `o` but `match(o.field) { V1 => here }`
+  LazyStructLoadInfo load_info;     // instructions, which fields to load, which to skip, etc.
+
+  AuxData_LazyObjectLoadFields(LocalVarPtr var_ref, TypePtr union_variant, StructFieldPtr field_ref, LazyStructLoadInfo load_info)
+    : var_ref(var_ref), union_variant(union_variant), field_ref(field_ref), load_info(std::move(load_info)) {
+  }
+};
+
+// AuxData_LazyMatchForUnion wraps `match(lazy_var)` or its field
+struct AuxData_LazyMatchForUnion final : ASTAuxData {
+  LocalVarPtr var_ref;              // comes from `lazy`
+  StructFieldPtr field_ref;         // not `match(o)`, but `match(o.field)`
+
+  AuxData_LazyMatchForUnion(LocalVarPtr var_ref, StructFieldPtr field_ref)
+    : var_ref(var_ref), field_ref(field_ref) {
   }
 };