--- a/tolk/fwd-declarations.h
+++ b/tolk/fwd-declarations.h
@@ -60,4 +60,6 @@ enum class FunctionInlineMode {
   noInline,
 };
 
+typedef int var_idx_t;
+
 } // namespace tolk