… tolk-tester/tests/a-tests.tolk
… tolk-tester/tests/allow-post-modification.tolk
… tolk-tester/tests/asm-arg-order.tolk
… tolk-tester/tests/assignment-tests.tolk
… tolk-tester/tests/bit-operators.tolk
… tolk-tester/tests/build-addr-tests.tolk
… tolk-tester/tests/calls-tests.tolk
… tolk-tester/tests/cells-slices.tolk
… tolk-tester/tests/codegen-check-demo.tolk
… tolk-tester/tests/comments-tests.tolk
… tolk-tester/tests/constants-tests.tolk
… tolk-tester/tests/dicts-demo.tolk
… tolk-tester/tests/generics-1.tolk
… tolk-tester/tests/generics-2.tolk
… tolk-tester/tests/generics-3.tolk
… tolk-tester/tests/generics-4.tolk
… tolk-tester/tests/handle-msg-1.tolk
… tolk-tester/tests/handle-msg-2.tolk
… tolk-tester/tests/handle-msg-3.tolk
… tolk-tester/tests/handle-msg-4.tolk
… tolk-tester/tests/handle-msg-5.tolk
… tolk-tester/tests/handle-msg-6.tolk
… tolk-tester/tests/handle-msg-7.tolk
… tolk-tester/tests/if-else-tests.tolk
… tolk-tester/tests/imports-tests.tolk
… tolk-tester/tests/imports/has-tinf-err.tolk
… tolk-tester/tests/imports/some-helpers.tolk
… tolk-tester/tests/imports/some-storage.tolk
… tolk-tester/tests/imports/use-dicts.ext.tolk
… tolk-tester/tests/indexed-access.tolk