@kambyzsahhamt-lang 2025-09-24
Transaction 0.32 then 0.97 then 0.94. This transaction was done. I played the game for $200 but it didn't pay. The amount of the game is my daughter has cancer and I need to buy medicine. UQDP-paXoqSb3BvRz2cq0V3QqbiXIC5FwAd0ZuKtKGVwKO7S      😭😭😭😭😭😭