@kambyzsahhamt-lang 2025-09-24
I don't have to buy medicine, for Christ's sake, I'm tired. The world is so worthless.Ruthless. I begged on Twitter for 25 days, but no one helped me. UQDP-paXoqSb3BvRz2cq0V3QqbiXIC5FwAd0ZuKtKGVwKO7S😭😭😭😭😭😭😭😭😭😭😭